% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add}
\alias{add}
\title{add}
\usage{
add(x)
}
\arguments{
\item{x}{a list to be summed. Useful for doing elementwise summation of a list of matrices.}
}
\value{
returns a single summed object (e.g., a matrix)
}
\description{
add
}
\examples{
x<-list(matrix(c(1:4),nrow=2),matrix(c(1:4),nrow=2))
add(x)
}
