% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajack.R
\name{ajack}
\alias{ajack}
\title{ajack}
\usage{
ajack(data, FUN, ...)
}
\arguments{
\item{data}{data to get the bias parameter (a) for}

\item{FUN}{a function to be applied to the data}

\item{...}{additional arguments passed to FUN}
}
\value{
a vector of accelleration parameters for use in BCa bootstrap intervals
}
\description{
ajack
}
\examples{
data<-DFSimulated()
ajack(data,DFanalysis, betasonly=TRUE, robust=FALSE)
}
