% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllBootResults.R
\name{AllBootResults}
\alias{AllBootResults}
\title{AllBootResults}
\usage{
AllBootResults(boot, lower = 0.025, upper = 0.975, data, FUN, ...)
}
\arguments{
\item{boot}{A matrix of bootstrap results}

\item{lower}{the lower alpha}

\item{upper}{the upper alpha}

\item{data}{the data used for analysis}

\item{FUN}{the function used for analysis}

\item{...}{additional arguments to pass to FUN}
}
\value{
a matrix of results. Includes the baseline results, all output from standardBootIntervals, all results from BCa for both the jackknife and bootstrap accelleration methods. The bootstrap accelleration method is experimental.
}
\description{
AllBootResults
}
\examples{
data<-DFSimulated()
boots<-NaiveBoot(data, groups="Rs", keepgroups=TRUE)
boots<-bootAnalysis(boots, cbind, DFanalysis, 1,2,3, robust=FALSE)
AllBootResults(boots, .025,.975, data, DFanalysis, 1,2,3, robust=FALSE)
}
