% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omisc_annotateggplot.R
\name{calculate_median}
\alias{calculate_median}
\title{calculate_median}
\usage{
p <- p + stat_summary(fun.data = calculate_median, geom = "text")

data.frame() retains the median value as numeric.
However, c() makes both y and label a character since paste0 produces an output of type character.
}
\arguments{
\item{x}{a vector of numeric values}

\item{round_to}{the number of decimal places to which the median needs to be rounded. Default is 4.}
}
\value{
A y coordinate and the median of the values in the vector supplied to the function. The median is rounded
to four decimal places by default, but this can be changed in the arguments. For example:
   y    label
67.5568 n = 100
These annotations are inserted into the graph object generated by ggplot at the coordinates specified under "y"
once this function is run as specified under usage
}
\description{
A function to annotate the graph with the median value of the observations.
}
\examples{
x <- runif(n = 100, min = 2, max = 200)
calculate_median(x)
}
\keyword{annotations}
\keyword{graph}
\keyword{median,}
\keyword{statistics,}
\keyword{summary}
