% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omisc_annotateggplot.R
\name{calculate_n}
\alias{calculate_n}
\title{calculate_median}
\usage{
p <- p + stat_summary(fun.data = calculate_n, geom = "text")

data.frame() retains the median value as numeric.
However, c() makes both y and label a character since paste0 produces an output of type character.
}
\arguments{
\item{x}{a vector of numeric values}
}
\value{
A y coordinate and the number of the values in the vector supplied to the function. The result is expressed
as a string "n = numeric value`. For example:
   y    label
67.5568 n = 100
These annotations are inserted into the graph object generated by ggplot at the coordinates specified under "y"
once this function is run as specified under usage
}
\description{
A function to annotate the graph with the number of samples
}
\examples{
x <- runif(n = 100, min = 2, max = 200)
calculate_n(x)
}
\keyword{annotations}
\keyword{graph}
\keyword{sample}
\keyword{size,}
