% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFSimulatedChisqnew.R
\name{DFSimulatedChisqNew}
\alias{DFSimulatedChisqNew}
\title{DFSimulatedChisqNew}
\usage{
DFSimulatedChisqNew(MZ = 250, DZ = 250, a2 = 0.3, c2 = 0.3, df = 5)
}
\arguments{
\item{MZ}{Number of MZ twins to simulate}

\item{DZ}{Number of DZ twins to simulate}

\item{a2}{Heritability (proportion of variance)}

\item{c2}{Shared environment (proportion of variance)}

\item{df}{Total degrees of freedom for the Chi-Square variable}
}
\value{
A dataframe of Chi-Square distributed outcome observations for MZ and DZ twins
}
\description{
DFSimulatedChisqNew
}
\examples{
TwinData<-DFSimulatedChisqNew(200,200,.3,.3, 10)
}
