% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubleEnter.R
\name{doubleEnter}
\alias{doubleEnter}
\title{DoubleEnter}
\usage{
doubleEnter(proband, sibling, Rs)
}
\arguments{
\item{proband}{The proband scores}

\item{sibling}{The matched sibling scores}

\item{Rs}{The relatedness coefficients}
}
\value{
A dataframe
}
\description{
DoubleEnter
}
\examples{
X<-DFSimulated(10,10,.2,.2)
Y<-doubleEnter(X[,"proband"], X[,"sibling"], X[,"Rs"])
}
