% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HoffPseudoStandard.R
\name{HoffPseudoStandard}
\alias{HoffPseudoStandard}
\title{HoffPseudoStandard}
\usage{
HoffPseudoStandard(betas, SDX, interceptvar)
}
\arguments{
\item{betas}{A vector of betas from a multilevel model}

\item{SDX}{A vector of the standard deviations of the X value for each of the X's associated with the bets}

\item{interceptvar}{A vector of the intercept variances at the level associated with the betas}
}
\value{
A vector of pseudostandardized coefficients
}
\description{
HoffPseudoStandard
}
\examples{
print("none")
}
