% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jackknife.R
\name{jackknife}
\alias{jackknife}
\title{jackknife}
\usage{
jackknife(data)
}
\arguments{
\item{data}{The data to jackknife}
}
\value{
a list of jackknife datasets
}
\description{
jackknife
}
\examples{
data<-cbind(1:10,1:10)
result<-jackknife(data)
lapply(result,mean)

}
