% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/justBetas.R
\name{justBetas}
\alias{justBetas}
\title{justBetas}
\usage{
justBetas(data, Y, X)
}
\arguments{
\item{data}{A data frame}

\item{Y}{The name or column number of the Y variable}

\item{X}{The name(s) or column number(s) of the X variables}
}
\value{
A vector of unstandardized beta weights
}
\description{
justBetas
}
\examples{
X<-stats::rnorm(100)
Y<-stats::rnorm(100)+5*(X)
data<-cbind(Y,X)
justBetas(data,1,2)
#if you want an intercept
Y<-stats::rnorm(100)+5*(X)+5
data<-cbind(Y,X,1)
justBetas(data,1,c(2:3))
}
