% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbind.R
\name{lbind}
\alias{lbind}
\title{lbind}
\usage{
lbind(index, alist, n)
}
\arguments{
\item{index}{a list of indexes. This should count the number of items to return in the final list}

\item{alist}{a list of objects to be passed to rbind. They should be grouped according to which objects will be combined (e.g., if 1,2,3 are to be passed to cbind then they should be adjacent to eachother).}

\item{n}{The number of objects in each group. Currently each group must consist of the same number of objects.}
}
\value{
a list
}
\description{
lbind is meant to be used in conjuction with lapply to combine elements of lists using rbind.
}
\examples{
alist<-list(c(1,1),c(2,2),c(3,3))
index<-list(1)
n<-3
lapply(index,lbind,alist,3)
}
