% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leave1out.R
\name{leave1out}
\alias{leave1out}
\title{leave1out}
\usage{
leave1out(x, data)
}
\arguments{
\item{x}{Which row(s) of data to leave out}

\item{data}{A dataframe or matrix.}
}
\value{
The reduced dataframe or matrix
}
\description{
leave1out
}
\examples{
data<-cbind(1:10,1:10)
leave1out(5,data)

}
