% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NaiveBoot.R
\name{NaiveBoot}
\alias{NaiveBoot}
\title{The Naive Bootstrap}
\usage{
NaiveBoot(data, B = 1000, groups = NULL, keepgroups = F, size = 1)
}
\arguments{
\item{data}{data to be bootstrapped}

\item{B}{number of bootstrap samples to take}

\item{groups}{grouping variable if there is one}

\item{keepgroups}{keep the grouping variable?}

\item{size}{size of the bootstrap resamples relative to the original sample}
}
\value{
a list of bootstrap resamples
}
\description{
The Naive Bootstrap
}
\examples{
X<-TestData()
Y<-NaiveBoot(X)
}
