% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omisc-package.R
\docType{package}
\name{omisc-package}
\alias{omisc}
\alias{omisc-package}
\title{omisc: Miscelaneous Tools For The Analysis Of Omics Data}
\description{
A set of miscelaneous tools to ease the analysis of omics data
  like the one obtained from microarray and bulk RNA-seq.
}
\author{
\strong{Maintainer}: Diego Diez \email{diego10ruiz@gmail.com} (\href{https://orcid.org/0000-0002-2325-4893}{ORCID})

}
\keyword{internal}
