% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_correlation}
\alias{plot_correlation}
\title{plot_correlation}
\usage{
plot_correlation(x, ...)
}
\arguments{
\item{x}{an R object.}

\item{...}{arguments passed down to methods.}

\item{log}{whether to use log CPMs.}
}
\description{
Correlation heatmap from samples.
}
