% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_enrichr}
\alias{plot_enrichr}
\title{plot_enrichr}
\usage{
plot_enrichr(x, db)
}
\arguments{
\item{x}{enrichR result.}
}
\description{
Plots the results of enrichR with some different choice of colors.
}
