% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene.R
\name{plot_gene}
\alias{plot_gene}
\title{Basic gene ploting function}
\usage{
plot_gene(
  symbol,
  genome,
  tracks = NULL,
  add.data = NULL,
  add.ann = NULL,
  add.ideogram = TRUE,
  from = NULL,
  to = NULL,
  biomart = NULL,
  ylim = NULL,
  ...
)
}
\arguments{
\item{symbol}{symbol of the gene to plot. Will be used to search using BiomartGeneRegionTrack.}

\item{genome}{genome assembly (e.g. mm10 or hg38).}

\item{tracks}{additional tracks to add.}

\item{add.data}{list with data to be added as a DataTrack.}

\item{add.ann}{annotation information.}

\item{add.ideogram}{whether to add an IdeogramTrack (FALSE by default to spead up testing).}

\item{from}{start genomic coordinates.}

\item{to}{end genomic coordinates.}

\item{biomart}{mart object (optional).}

\item{ylim}{sets the limits for y-axis (default: NULL).}

\item{...}{further arguments passed to plotTracks.}
}
\value{
nothing but produces a plot as side effect.
}
\description{
Plots a gene from a specified assembly using Gviz package. Optionaly plots some data (e.g. read alignments). Enables zooming.
}
\examples{
NULL
}
