% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_heatmap}
\alias{plot_heatmap}
\alias{plot_heatmap.DGEList}
\alias{plot_heatmap.matrix}
\title{plot_heatmap}
\usage{
plot_heatmap(x, ...)

\method{plot_heatmap}{DGEList}(
  x,
  log = TRUE,
  batch = NULL,
  design = NULL,
  symbol.col = "SYMBOL",
  top_ann = NULL,
  top_ann_col = NULL,
  ...
)

\method{plot_heatmap}{matrix}(x, scale = TRUE, show_column_names = TRUE, ...)
}
\arguments{
\item{x}{object to plot.}

\item{...}{}

\item{log}{whether to use log CPMs.}

\item{batch}{batch variable for removeBatchEffect.}

\item{design}{design matrix for removeBatchEffect.}

\item{symbol.col}{column name with gene symbols.}

\item{top_ann}{names of columns to be used as top annotations.}

\item{top_ann_col}{color definition for the categories in the top annotations.}

\item{show_column_names}{whether to show column names (default: TRUE).}
}
\description{
plot_heatmap
}
