% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_hist}
\alias{plot_hist}
\alias{plot_hist.MArrayLM}
\title{plot_hist}
\usage{
plot_hist(x, ...)

\method{plot_hist}{MArrayLM}(x, coef = NULL)
}
\arguments{
\item{x}{an R object.}

\item{...}{arguments passed down to methods.}

\item{coef}{name of coefficient.}
}
\value{

}
\description{
Plots histogram from various types of objects.
}
