% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strand.R
\name{plot_strand_qc}
\alias{plot_strand_qc}
\title{plot_strand_qc}
\usage{
plot_strand_qc(x)
}
\arguments{
\item{x}{a data.frame like object obtained with read_star().}
}
\description{
Plots to facilitate identifying the strandness of the sequencing data.
}
