% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venn.R
\name{plot_venn}
\alias{plot_venn}
\alias{plot_venn.matrix}
\alias{plot_venn.data.frame}
\alias{plot_venn.list}
\title{plotVenn}
\usage{
plot_venn(x, ...)

\method{plot_venn}{matrix}(
  x,
  add.universe = FALSE,
  euler = FALSE,
  scaled = FALSE,
  filename = NULL,
  fontfamily = "sans",
  cat.fontfamily = "sans",
  main.fontfamily = "sans",
  ...
)

\method{plot_venn}{data.frame}(
  x,
  euler = FALSE,
  scaled = FALSE,
  filename = NULL,
  fontfamily = "sans",
  cat.fontfamily = "sans",
  main.fontfamily = "sans",
  ...
)

\method{plot_venn}{list}(
  x,
  euler = FALSE,
  scaled = FALSE,
  filename = NULL,
  fontfamily = "sans",
  cat.fontfamily = "sans",
  main.fontfamily = "sans",
  ...
)
}
\arguments{
\item{x}{matrix of values (similar to what is fed to VennDiagram in limma package).}

\item{...}{additional parameters passed down to venn.diagrama()}

\item{add.universe}{logical; whether to add a group containing all elements.}
}
\description{
Plot a venn diagram with VennDiagram package using some nice defaults
}
\examples{
NULL
}
