% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_volcano}
\alias{plot_volcano}
\alias{plot_volcano.MArrayLM}
\alias{plot_correlation.DGEList}
\title{plot_volcano}
\usage{
plot_volcano(x, ...)

\method{plot_volcano}{MArrayLM}(
  x,
  coef = 1,
  top_genes = NULL,
  lfc = 1,
  fdr = 0.01,
  use.column = "symbol",
  color.by = NULL
)

\method{plot_correlation}{DGEList}(x, log = TRUE, ...)
}
\arguments{
\item{x}{an R object.}

\item{...}{arguments passed down to methods.}

\item{coef}{name of the coefficient.}

\item{top_genes}{number of genes to highlight.}

\item{lfc}{logFC cutoff for top genes.}

\item{fdr}{FDR cutoff for top genes.}

\item{use.column}{column containing gene names (default: symbol).}

\item{color.by}{column to color points.}
}
\description{
Volcano plot from various types of objects.
}
