% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voom.R
\name{plot_voom}
\alias{plot_voom}
\alias{plot_voom.EList}
\title{Plots the mean-variance trend.}
\usage{
plot_voom(x)

\method{plot_voom}{EList}(x)
}
\arguments{
\item{x}{Output of voom() function.}
}
\description{
This function uses the information stored in the object returned by voom()
when called with save.plot = TRUE.
}
