% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{resample}
\alias{resample}
\title{resample}
\usage{
resample(X, size)
}
\arguments{
\item{X}{A vector to be resamples}

\item{size}{The size of the resulting vector. Should be a number such that size*nrow(X) is a whole number}
}
\value{
A vector of resampled X values
}
\description{
resample
}
\examples{
X<-c(1:10)
resample(X,.5)
}
