% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RK.R
\name{RK}
\alias{RK}
\title{RK}
\usage{
RK(proband, sibling, Rs, DE = T)
}
\arguments{
\item{proband}{column name or number of the proband}

\item{sibling}{column name or number of the siblings}

\item{Rs}{column name or number of the relatedness coefficients}

\item{DE}{Should the data be double entered?}
}
\value{
A dataframe
}
\description{
RK
}
\examples{
X<-DFSimulated(100,100,.3,.3)
Y<-RK(X[,1],X[,2],X[,3])
}
