% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestData.R
\name{TestData}
\alias{TestData}
\title{TestData}
\usage{
TestData(nobs = 1000, intercept = 0, beta = 5, meanX = 0, sdX = 1, sdYerr = 1)
}
\arguments{
\item{nobs}{Number of observations, defaults to 1000}

\item{intercept}{Intercept of the regression. Defaults to 0}

\item{beta}{Beta for the regression equation, defaults to 5}

\item{meanX}{Mean of X, defaults to 0}

\item{sdX}{Standard deviation of X, defaults to 1}

\item{sdYerr}{Variance of the error term of Y, defaults to 1}
}
\value{
A dataframe with an X and Y variable produced by the entered parameters
}
\description{
Simple function for creating a dataset of two related variables.
}
\examples{
X<-TestData()
}
