% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{to_tidy}
\alias{to_tidy}
\alias{to_tidy.matrix}
\alias{to_tidy.data.frame}
\alias{to_tidy.tbl_df}
\title{to_tidy}
\usage{
to_tidy(x, ...)

\method{to_tidy}{matrix}(x, row.name = "row", ...)

\method{to_tidy}{data.frame}(x, row.name = "row", ...)

\method{to_tidy}{tbl_df}(
  x,
  row.name = "row",
  col.name = "col",
  value.name = "value",
  stringsAsFactors = FALSE,
  ...
)
}
\arguments{
\item{x}{a matrix, data.frame or tibble object.}

\item{...}{arguments passed to methods.}

\item{row.name}{name for row data.}

\item{col.name}{name for column data.}

\item{value.name}{name use for value column.}

\item{stringsAsFactors}{logical; whether to convert col/row names to factors (preserving original ordering).}
}
\description{
Convert a matrix, data.frame or tibble into a tidy tibble.
}
