% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unibootsample.R
\name{unibootsample}
\alias{unibootsample}
\title{unibootsample}
\usage{
unibootsample(data, size)
}
\arguments{
\item{data}{A dataframe or matrix to be univariately bootstrapped}

\item{size}{size of each bootstrap sample as a fraction of the total sample size. Total sample size must be evenly divisible by "size".}
}
\value{
A matrix or dataframe with nrow=nrow(X)*size
}
\description{
unibootsample
}
\examples{
X<-c(0:9)
Y<-c(20:29)
Z<-cbind(X,Y)
unibootsample(Z,1)
}
