% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unibootVar.R
\name{unibootVar}
\alias{unibootVar}
\title{unibootVar}
\usage{
unibootVar(X, times)
}
\arguments{
\item{X}{The variable}

\item{times}{The number of times the variable is repeated in the univariate sampling frame. This is going to be equal to the number of variables being univariately sampled}
}
\value{
The variance of the variable in the univariate sampling frame
}
\description{
unibootVar
}
\examples{
X<-c(1,2)
times<-100
unibootVar(X,times)
var(X)
}
