% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zScore.R
\name{zScore}
\alias{zScore}
\title{zScore}
\usage{
zScore(X, times)
}
\arguments{
\item{X}{vector to be converted to z scores}

\item{times}{exponent controlling the denominator scaling}
}
\value{
Returns a vector of z scores
}
\description{
zScore
}
\examples{
X<-c(1:10)
zScore(X, times=1)
}
