% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zScoreData.R
\name{zScoreData}
\alias{zScoreData}
\title{centerData}
\usage{
zScoreData(data)
}
\arguments{
\item{data}{The data to be converted to z scores}
}
\value{
Data converted to z scores
}
\description{
centerData
}
\examples{
X<-data.frame(X=c(1:4),Y=c(6:9))
zScoreData(X)
}
