% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_ma}
\alias{plot_ma}
\alias{plot_ma.MArrayLM}
\title{plot_ma}
\usage{
plot_ma(x, ...)

\method{plot_ma}{MArrayLM}(x, coef = 1, color.by = NULL)
}
\arguments{
\item{x}{an R object.}

\item{...}{arguments passed down to methods.}

\item{coef}{name of the coefficient.}

\item{color.by}{column to color points.}
}
\description{
MA plot from various types of objects.
}
