% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zScore.R
\name{zScore}
\alias{zScore}
\title{Title}
\usage{
zScore(X, reps = 1)
}
\arguments{
\item{X}{The vector to be turned into z scores}

\item{reps}{The number of reps the vector is to be repeated. This will only be used in univariate bootstrapping. The default is 1.}
}
\value{
A vector of z scores.
}
\description{
Title
}
\examples{
X<-c(1:10)
zScore(X)
}
